Lucas reservations

+++ database/migrations/2014_10_12_000000_create_users_table.php
@@ -22,7 +22,7 @@ class CreateUsersTable extends Migration
             $table->string('password');
             $table->string('avatar')->nullable();
             $table->date('date_naissance');
-            $table->integer('note_eval')->nullable();
+            $table->double('note_eval', 2, 1)->nullable();
             $table->enum('role', ['user', 'gerant'])->default('user');
             $table->boolean('signale')->default(false);
             $table->rememberToken();
